api/requirements.txt
uvicorn==0.17.6
fastapi==0.78.0
pytz
pymongo==4.6.1
pydantic==1.8.2
python-dateutil==2.8.2
